Struct aws_sdk_emr::client::fluent_builders::RunJobFlow

pub struct RunJobFlow<C = DynConnector, M = AwsMiddleware, R = Standard> { /* fields omitted */ }

Fluent builder constructing a request to RunJobFlow.

RunJobFlow creates and starts running a new cluster (job flow). The cluster runs the steps specified. After the steps complete, the cluster stops and the HDFS partition is lost. To prevent loss of data, configure the last step of the job flow to store results in Amazon S3. If the JobFlowInstancesConfig KeepJobFlowAliveWhenNoSteps parameter is set to TRUE, the cluster transitions to the WAITING state rather than shutting down after the steps have completed.

For additional protection, you can set the JobFlowInstancesConfig TerminationProtected parameter to TRUE to lock the cluster and prevent it from being terminated by API call, user intervention, or in the event of a job flow error.

A maximum of 256 steps are allowed in each job flow.

If your cluster is long-running (such as a Hive data warehouse) or complex, you may require more than 256 steps to process your data. You can bypass the 256-step limitation in various ways, including using the SSH shell to connect to the master node and submitting queries directly to the software running on the master node, such as Hive and Hadoop. For more information on how to do this, see Add More than 256 Steps to a Cluster in the Amazon EMR Management Guide.

For long running clusters, we recommend that you periodically store your results.

The instance fleets configuration is available only in Amazon EMR versions 4.8.0 and later, excluding 5.0.x versions. The RunJobFlow request can contain InstanceFleets parameters or InstanceGroups parameters, but not both.

Implementations

impl<C, M, R> RunJobFlow<C, M, R> where
C: SmithyConnector,
M: SmithyMiddleware<C>,
R: NewRequestPolicy,

pub async fn send(self) -> Result<RunJobFlowOutput, SdkError<RunJobFlowError>> where
R::Policy: SmithyRetryPolicy<RunJobFlowInputOperationOutputAlias, RunJobFlowOutput, RunJobFlowError, RunJobFlowInputOperationRetryAlias>,

Sends the request and returns the response.

If an error occurs, an SdkError will be returned with additional details that can be matched against.

By default, any retryable failures will be retried twice. Retry behavior is configurable with the RetryConfig, which can be set when configuring the client.

pub fn name(self, inp: impl Into<String>) -> Self

The name of the job flow.

pub fn set_name(self, input: Option<String>) -> Self

The name of the job flow.

pub fn log_uri(self, inp: impl Into<String>) -> Self

The location in Amazon S3 to write the log files of the job flow. If a value is not provided, logs are not created.

pub fn set_log_uri(self, input: Option<String>) -> Self

The location in Amazon S3 to write the log files of the job flow. If a value is not provided, logs are not created.

pub fn log_encryption_kms_key_id(self, inp: impl Into<String>) -> Self

The KMS key used for encrypting log files. If a value is not provided, the logs remain encrypted by AES-256. This attribute is only available with Amazon EMR version 5.30.0 and later, excluding Amazon EMR 6.0.0.

pub fn set_log_encryption_kms_key_id(self, input: Option<String>) -> Self

The KMS key used for encrypting log files. If a value is not provided, the logs remain encrypted by AES-256. This attribute is only available with Amazon EMR version 5.30.0 and later, excluding Amazon EMR 6.0.0.

pub fn additional_info(self, inp: impl Into<String>) -> Self

A JSON string for selecting additional features.

pub fn set_additional_info(self, input: Option<String>) -> Self

A JSON string for selecting additional features.

pub fn ami_version(self, inp: impl Into<String>) -> Self

Applies only to Amazon EMR AMI versions 3.x and 2.x. For Amazon EMR releases 4.0 and later, ReleaseLabel is used. To specify a custom AMI, use CustomAmiID.

pub fn set_ami_version(self, input: Option<String>) -> Self

Applies only to Amazon EMR AMI versions 3.x and 2.x. For Amazon EMR releases 4.0 and later, ReleaseLabel is used. To specify a custom AMI, use CustomAmiID.

pub fn release_label(self, inp: impl Into<String>) -> Self

The Amazon EMR release label, which determines the version of open-source application packages installed on the cluster. Release labels are in the form emr-x.x.x, where x.x.x is an Amazon EMR release version such as emr-5.14.0. For more information about Amazon EMR release versions and included application versions and features, see https://docs.aws.amazon.com/emr/latest/ReleaseGuide/. The release label applies only to Amazon EMR releases version 4.0 and later. Earlier versions use AmiVersion.

pub fn set_release_label(self, input: Option<String>) -> Self

The Amazon EMR release label, which determines the version of open-source application packages installed on the cluster. Release labels are in the form emr-x.x.x, where x.x.x is an Amazon EMR release version such as emr-5.14.0. For more information about Amazon EMR release versions and included application versions and features, see https://docs.aws.amazon.com/emr/latest/ReleaseGuide/. The release label applies only to Amazon EMR releases version 4.0 and later. Earlier versions use AmiVersion.

pub fn instances(self, inp: JobFlowInstancesConfig) -> Self

A specification of the number and type of Amazon EC2 instances.

pub fn set_instances(self, input: Option<JobFlowInstancesConfig>) -> Self

A specification of the number and type of Amazon EC2 instances.

pub fn steps(self, inp: impl Into<StepConfig>) -> Self

Appends an item to Steps.

To override the contents of this collection use set_steps.

A list of steps to run.

pub fn set_steps(self, input: Option<Vec<StepConfig>>) -> Self

A list of steps to run.

pub fn bootstrap_actions(self, inp: impl Into<BootstrapActionConfig>) -> Self

Appends an item to BootstrapActions.

To override the contents of this collection use set_bootstrap_actions.

A list of bootstrap actions to run before Hadoop starts on the cluster nodes.

pub fn set_bootstrap_actions(
self,
input: Option<Vec<BootstrapActionConfig>>
) -> Self

A list of bootstrap actions to run before Hadoop starts on the cluster nodes.

pub fn supported_products(self, inp: impl Into<String>) -> Self

Appends an item to SupportedProducts.

To override the contents of this collection use set_supported_products.

For Amazon EMR releases 3.x and 2.x. For Amazon EMR releases 4.x and later, use Applications.

A list of strings that indicates third-party software to use. For more information, see the Amazon EMR Developer Guide. Currently supported values are:

"mapr-m3" - launch the job flow using MapR M3 Edition.

"mapr-m5" - launch the job flow using MapR M5 Edition.

pub fn set_supported_products(self, input: Option<Vec<String>>) -> Self

For Amazon EMR releases 3.x and 2.x. For Amazon EMR releases 4.x and later, use Applications.

A list of strings that indicates third-party software to use. For more information, see the Amazon EMR Developer Guide. Currently supported values are:

"mapr-m3" - launch the job flow using MapR M3 Edition.

"mapr-m5" - launch the job flow using MapR M5 Edition.

pub fn new_supported_products(
self,
inp: impl Into<SupportedProductConfig>
) -> Self

Appends an item to NewSupportedProducts.

To override the contents of this collection use set_new_supported_products.

For Amazon EMR releases 3.x and 2.x. For Amazon EMR releases 4.x and later, use Applications.

A list of strings that indicates third-party software to use with the job flow that accepts a user argument list. EMR accepts and forwards the argument list to the corresponding installation script as bootstrap action arguments. For more information, see "Launch a Job Flow on the MapR Distribution for Hadoop" in the Amazon EMR Developer Guide. Supported values are:

"mapr-m3" - launch the cluster using MapR M3 Edition.

"mapr-m5" - launch the cluster using MapR M5 Edition.

"mapr" with the user arguments specifying "--edition,m3" or "--edition,m5" - launch the job flow using MapR M3 or M5 Edition respectively.

"mapr-m7" - launch the cluster using MapR M7 Edition.

"hunk" - launch the cluster with the Hunk Big Data Analytics Platform.

"hue"- launch the cluster with Hue installed.

"spark" - launch the cluster with Apache Spark installed.

"ganglia" - launch the cluster with the Ganglia Monitoring System installed.

pub fn set_new_supported_products(
self,
input: Option<Vec<SupportedProductConfig>>
) -> Self

For Amazon EMR releases 3.x and 2.x. For Amazon EMR releases 4.x and later, use Applications.

A list of strings that indicates third-party software to use with the job flow that accepts a user argument list. EMR accepts and forwards the argument list to the corresponding installation script as bootstrap action arguments. For more information, see "Launch a Job Flow on the MapR Distribution for Hadoop" in the Amazon EMR Developer Guide. Supported values are:

"mapr-m3" - launch the cluster using MapR M3 Edition.

"mapr-m5" - launch the cluster using MapR M5 Edition.

"mapr" with the user arguments specifying "--edition,m3" or "--edition,m5" - launch the job flow using MapR M3 or M5 Edition respectively.

"mapr-m7" - launch the cluster using MapR M7 Edition.

"hunk" - launch the cluster with the Hunk Big Data Analytics Platform.

"hue"- launch the cluster with Hue installed.

"spark" - launch the cluster with Apache Spark installed.

"ganglia" - launch the cluster with the Ganglia Monitoring System installed.

pub fn applications(self, inp: impl Into<Application>) -> Self

Appends an item to Applications.

To override the contents of this collection use set_applications.

Applies to Amazon EMR releases 4.0 and later. A case-insensitive list of applications for Amazon EMR to install and configure when launching the cluster. For a list of applications available for each Amazon EMR release version, see the Amazon EMR Release Guide.

pub fn set_applications(self, input: Option<Vec<Application>>) -> Self

Applies to Amazon EMR releases 4.0 and later. A case-insensitive list of applications for Amazon EMR to install and configure when launching the cluster. For a list of applications available for each Amazon EMR release version, see the Amazon EMR Release Guide.

pub fn configurations(self, inp: impl Into<Configuration>) -> Self

Appends an item to Configurations.

To override the contents of this collection use set_configurations.

For Amazon EMR releases 4.0 and later. The list of configurations supplied for the EMR cluster you are creating.

pub fn set_configurations(self, input: Option<Vec<Configuration>>) -> Self

For Amazon EMR releases 4.0 and later. The list of configurations supplied for the EMR cluster you are creating.

pub fn visible_to_all_users(self, inp: bool) -> Self

Set this value to true so that IAM principals in the Amazon Web Services account associated with the cluster can perform EMR actions on the cluster that their IAM policies allow. This value defaults to true for clusters created using the EMR API or the CLI create-cluster command.

When set to false, only the IAM principal that created the cluster and the Amazon Web Services account root user can perform EMR actions for the cluster, regardless of the IAM permissions policies attached to other IAM principals. For more information, see Understanding the EMR Cluster VisibleToAllUsers Setting in the Amazon EMRManagement Guide.

pub fn set_visible_to_all_users(self, input: Option<bool>) -> Self

Set this value to true so that IAM principals in the Amazon Web Services account associated with the cluster can perform EMR actions on the cluster that their IAM policies allow. This value defaults to true for clusters created using the EMR API or the CLI create-cluster command.

When set to false, only the IAM principal that created the cluster and the Amazon Web Services account root user can perform EMR actions for the cluster, regardless of the IAM permissions policies attached to other IAM principals. For more information, see Understanding the EMR Cluster VisibleToAllUsers Setting in the Amazon EMRManagement Guide.

pub fn job_flow_role(self, inp: impl Into<String>) -> Self

Also called instance profile and EC2 role. An IAM role for an EMR cluster. The EC2 instances of the cluster assume this role. The default role is EMR_EC2_DefaultRole. In order to use the default role, you must have already created it using the CLI or console.

pub fn set_job_flow_role(self, input: Option<String>) -> Self

Also called instance profile and EC2 role. An IAM role for an EMR cluster. The EC2 instances of the cluster assume this role. The default role is EMR_EC2_DefaultRole. In order to use the default role, you must have already created it using the CLI or console.

pub fn service_role(self, inp: impl Into<String>) -> Self

The IAM role that Amazon EMR assumes in order to access Amazon Web Services resources on your behalf.

pub fn set_service_role(self, input: Option<String>) -> Self

The IAM role that Amazon EMR assumes in order to access Amazon Web Services resources on your behalf.

pub fn tags(self, inp: impl Into<Tag>) -> Self

Appends an item to Tags.

To override the contents of this collection use set_tags.

A list of tags to associate with a cluster and propagate to Amazon EC2 instances.

pub fn set_tags(self, input: Option<Vec<Tag>>) -> Self

A list of tags to associate with a cluster and propagate to Amazon EC2 instances.

pub fn security_configuration(self, inp: impl Into<String>) -> Self

The name of a security configuration to apply to the cluster.

pub fn set_security_configuration(self, input: Option<String>) -> Self

The name of a security configuration to apply to the cluster.

pub fn auto_scaling_role(self, inp: impl Into<String>) -> Self

An IAM role for automatic scaling policies. The default role is EMR_AutoScaling_DefaultRole. The IAM role provides permissions that the automatic scaling feature requires to launch and terminate EC2 instances in an instance group.

pub fn set_auto_scaling_role(self, input: Option<String>) -> Self

An IAM role for automatic scaling policies. The default role is EMR_AutoScaling_DefaultRole. The IAM role provides permissions that the automatic scaling feature requires to launch and terminate EC2 instances in an instance group.

pub fn scale_down_behavior(self, inp: ScaleDownBehavior) -> Self

Specifies the way that individual Amazon EC2 instances terminate when an automatic scale-in activity occurs or an instance group is resized. TERMINATE_AT_INSTANCE_HOUR indicates that Amazon EMR terminates nodes at the instance-hour boundary, regardless of when the request to terminate the instance was submitted. This option is only available with Amazon EMR 5.1.0 and later and is the default for clusters created using that version. TERMINATE_AT_TASK_COMPLETION indicates that Amazon EMR adds nodes to a deny list and drains tasks from nodes before terminating the Amazon EC2 instances, regardless of the instance-hour boundary. With either behavior, Amazon EMR removes the least active nodes first and blocks instance termination if it could lead to HDFS corruption. TERMINATE_AT_TASK_COMPLETION available only in Amazon EMR version 4.1.0 and later, and is the default for versions of Amazon EMR earlier than 5.1.0.

pub fn set_scale_down_behavior(self, input: Option<ScaleDownBehavior>) -> Self

Specifies the way that individual Amazon EC2 instances terminate when an automatic scale-in activity occurs or an instance group is resized. TERMINATE_AT_INSTANCE_HOUR indicates that Amazon EMR terminates nodes at the instance-hour boundary, regardless of when the request to terminate the instance was submitted. This option is only available with Amazon EMR 5.1.0 and later and is the default for clusters created using that version. TERMINATE_AT_TASK_COMPLETION indicates that Amazon EMR adds nodes to a deny list and drains tasks from nodes before terminating the Amazon EC2 instances, regardless of the instance-hour boundary. With either behavior, Amazon EMR removes the least active nodes first and blocks instance termination if it could lead to HDFS corruption. TERMINATE_AT_TASK_COMPLETION available only in Amazon EMR version 4.1.0 and later, and is the default for versions of Amazon EMR earlier than 5.1.0.

pub fn custom_ami_id(self, inp: impl Into<String>) -> Self

Available only in Amazon EMR version 5.7.0 and later. The ID of a custom Amazon EBS-backed Linux AMI. If specified, Amazon EMR uses this AMI when it launches cluster EC2 instances. For more information about custom AMIs in Amazon EMR, see Using a Custom AMI in the Amazon EMR Management Guide. If omitted, the cluster uses the base Linux AMI for the ReleaseLabel specified. For Amazon EMR versions 2.x and 3.x, use AmiVersion instead.

For information about creating a custom AMI, see Creating an Amazon EBS-Backed Linux AMI in the Amazon Elastic Compute Cloud User Guide for Linux Instances. For information about finding an AMI ID, see Finding a Linux AMI.

pub fn set_custom_ami_id(self, input: Option<String>) -> Self

Available only in Amazon EMR version 5.7.0 and later. The ID of a custom Amazon EBS-backed Linux AMI. If specified, Amazon EMR uses this AMI when it launches cluster EC2 instances. For more information about custom AMIs in Amazon EMR, see Using a Custom AMI in the Amazon EMR Management Guide. If omitted, the cluster uses the base Linux AMI for the ReleaseLabel specified. For Amazon EMR versions 2.x and 3.x, use AmiVersion instead.

For information about creating a custom AMI, see Creating an Amazon EBS-Backed Linux AMI in the Amazon Elastic Compute Cloud User Guide for Linux Instances. For information about finding an AMI ID, see Finding a Linux AMI.

pub fn ebs_root_volume_size(self, inp: i32) -> Self

The size, in GiB, of the Amazon EBS root device volume of the Linux AMI that is used for each EC2 instance. Available in Amazon EMR version 4.x and later.

pub fn set_ebs_root_volume_size(self, input: Option<i32>) -> Self

The size, in GiB, of the Amazon EBS root device volume of the Linux AMI that is used for each EC2 instance. Available in Amazon EMR version 4.x and later.

pub fn repo_upgrade_on_boot(self, inp: RepoUpgradeOnBoot) -> Self

Applies only when CustomAmiID is used. Specifies which updates from the Amazon Linux AMI package repositories to apply automatically when the instance boots using the AMI. If omitted, the default is SECURITY, which indicates that only security updates are applied. If NONE is specified, no updates are applied, and all updates must be applied manually.

pub fn set_repo_upgrade_on_boot(self, input: Option<RepoUpgradeOnBoot>) -> Self

Applies only when CustomAmiID is used. Specifies which updates from the Amazon Linux AMI package repositories to apply automatically when the instance boots using the AMI. If omitted, the default is SECURITY, which indicates that only security updates are applied. If NONE is specified, no updates are applied, and all updates must be applied manually.

pub fn kerberos_attributes(self, inp: KerberosAttributes) -> Self

Attributes for Kerberos configuration when Kerberos authentication is enabled using a security configuration. For more information see Use Kerberos Authentication in the Amazon EMR Management Guide.

pub fn set_kerberos_attributes(self, input: Option<KerberosAttributes>) -> Self

Attributes for Kerberos configuration when Kerberos authentication is enabled using a security configuration. For more information see Use Kerberos Authentication in the Amazon EMR Management Guide.

pub fn step_concurrency_level(self, inp: i32) -> Self

Specifies the number of steps that can be executed concurrently. The default value is 1. The maximum value is 256.

pub fn set_step_concurrency_level(self, input: Option<i32>) -> Self

Specifies the number of steps that can be executed concurrently. The default value is 1. The maximum value is 256.

pub fn managed_scaling_policy(self, inp: ManagedScalingPolicy) -> Self

The specified managed scaling policy for an Amazon EMR cluster.

pub fn set_managed_scaling_policy(
self,
input: Option<ManagedScalingPolicy>
) -> Self

The specified managed scaling policy for an Amazon EMR cluster.

pub fn placement_group_configs(
self,
inp: impl Into<PlacementGroupConfig>
) -> Self

Appends an item to PlacementGroupConfigs.

To override the contents of this collection use set_placement_group_configs.

The specified placement group configuration for an Amazon EMR cluster.

pub fn set_placement_group_configs(
self,
input: Option<Vec<PlacementGroupConfig>>
) -> Self

The specified placement group configuration for an Amazon EMR cluster.

pub fn auto_termination_policy(self, inp: AutoTerminationPolicy) -> Self

An auto-termination policy for an Amazon EMR cluster. An auto-termination policy defines the amount of idle time in seconds after which a cluster automatically terminates. For alternative cluster termination options, see Control cluster termination.

pub fn set_auto_termination_policy(
self,
input: Option<AutoTerminationPolicy>
) -> Self

An auto-termination policy for an Amazon EMR cluster. An auto-termination policy defines the amount of idle time in seconds after which a cluster automatically terminates. For alternative cluster termination options, see Control cluster termination.