Module aws_sdk_emr::error

Modules

internal_server_error

See InternalServerError

internal_server_exception

See InternalServerException

invalid_request_exception

See InvalidRequestException

Structs

AddInstanceFleetError

AddInstanceGroupsError

AddJobFlowStepsError

AddTagsError

CancelStepsError

CreateSecurityConfigurationError

CreateStudioError

CreateStudioSessionMappingError

DeleteSecurityConfigurationError

DeleteStudioError

DeleteStudioSessionMappingError

DescribeClusterError

DescribeJobFlowsError

DescribeNotebookExecutionError

DescribeReleaseLabelError

DescribeSecurityConfigurationError

DescribeStepError

DescribeStudioError

GetAutoTerminationPolicyError

GetBlockPublicAccessConfigurationError

GetManagedScalingPolicyError

GetStudioSessionMappingError

InternalServerError

Indicates that an error occurred while processing the request and that the request was

InternalServerException

This exception occurs when there is an internal failure in the Amazon EMR

InvalidRequestException

This exception occurs when there is something wrong with user input.

ListBootstrapActionsError

ListClustersError

ListInstanceFleetsError

ListInstanceGroupsError

ListInstancesError

ListNotebookExecutionsError

ListReleaseLabelsError

ListSecurityConfigurationsError

ListStepsError

ListStudioSessionMappingsError

ListStudiosError

ModifyClusterError

ModifyInstanceFleetError

ModifyInstanceGroupsError

PutAutoScalingPolicyError

PutAutoTerminationPolicyError

PutBlockPublicAccessConfigurationError

PutManagedScalingPolicyError

RemoveAutoScalingPolicyError

RemoveAutoTerminationPolicyError

RemoveManagedScalingPolicyError

RemoveTagsError

RunJobFlowError

SetTerminationProtectionError

SetVisibleToAllUsersError

StartNotebookExecutionError

StopNotebookExecutionError

TerminateJobFlowsError

UpdateStudioError

UpdateStudioSessionMappingError

Enums

AddInstanceFleetErrorKind

AddInstanceGroupsErrorKind

AddJobFlowStepsErrorKind

AddTagsErrorKind

CancelStepsErrorKind

CreateSecurityConfigurationErrorKind

CreateStudioErrorKind

CreateStudioSessionMappingErrorKind

DeleteSecurityConfigurationErrorKind

DeleteStudioErrorKind

DeleteStudioSessionMappingErrorKind

DescribeClusterErrorKind

DescribeJobFlowsErrorKind

DescribeNotebookExecutionErrorKind

DescribeReleaseLabelErrorKind

DescribeSecurityConfigurationErrorKind

DescribeStepErrorKind

DescribeStudioErrorKind

GetAutoTerminationPolicyErrorKind

GetBlockPublicAccessConfigurationErrorKind

GetManagedScalingPolicyErrorKind

GetStudioSessionMappingErrorKind

ListBootstrapActionsErrorKind

ListClustersErrorKind

ListInstanceFleetsErrorKind

ListInstanceGroupsErrorKind

ListInstancesErrorKind

ListNotebookExecutionsErrorKind

ListReleaseLabelsErrorKind

ListSecurityConfigurationsErrorKind

ListStepsErrorKind

ListStudioSessionMappingsErrorKind

ListStudiosErrorKind

ModifyClusterErrorKind

ModifyInstanceFleetErrorKind

ModifyInstanceGroupsErrorKind

PutAutoScalingPolicyErrorKind

PutAutoTerminationPolicyErrorKind

PutBlockPublicAccessConfigurationErrorKind

PutManagedScalingPolicyErrorKind

RemoveAutoScalingPolicyErrorKind

RemoveAutoTerminationPolicyErrorKind

RemoveManagedScalingPolicyErrorKind

RemoveTagsErrorKind

RunJobFlowErrorKind

SetTerminationProtectionErrorKind

SetVisibleToAllUsersErrorKind

StartNotebookExecutionErrorKind

StopNotebookExecutionErrorKind

TerminateJobFlowsErrorKind

UpdateStudioErrorKind

UpdateStudioSessionMappingErrorKind