Module aws_sdk_emr::input

Modules

add_instance_fleet_input

See AddInstanceFleetInput

add_instance_groups_input

See AddInstanceGroupsInput

add_job_flow_steps_input

See AddJobFlowStepsInput

add_tags_input

See AddTagsInput

cancel_steps_input

See CancelStepsInput

create_security_configuration_input

See CreateSecurityConfigurationInput

create_studio_input

See CreateStudioInput

create_studio_session_mapping_input

See CreateStudioSessionMappingInput

delete_security_configuration_input

See DeleteSecurityConfigurationInput

delete_studio_input

See DeleteStudioInput

delete_studio_session_mapping_input

See DeleteStudioSessionMappingInput

describe_cluster_input

See DescribeClusterInput

describe_job_flows_input

See DescribeJobFlowsInput

describe_notebook_execution_input

See DescribeNotebookExecutionInput

describe_release_label_input

See DescribeReleaseLabelInput

describe_security_configuration_input

See DescribeSecurityConfigurationInput

describe_step_input

See DescribeStepInput

describe_studio_input

See DescribeStudioInput

get_auto_termination_policy_input

See GetAutoTerminationPolicyInput

get_block_public_access_configuration_input

See GetBlockPublicAccessConfigurationInput

get_managed_scaling_policy_input

See GetManagedScalingPolicyInput

get_studio_session_mapping_input

See GetStudioSessionMappingInput

list_bootstrap_actions_input

See ListBootstrapActionsInput

list_clusters_input

See ListClustersInput

list_instance_fleets_input

See ListInstanceFleetsInput

list_instance_groups_input

See ListInstanceGroupsInput

list_instances_input

See ListInstancesInput

list_notebook_executions_input

See ListNotebookExecutionsInput

list_release_labels_input

See ListReleaseLabelsInput

list_security_configurations_input

See ListSecurityConfigurationsInput

list_steps_input

See ListStepsInput

list_studio_session_mappings_input

See ListStudioSessionMappingsInput

list_studios_input

See ListStudiosInput

modify_cluster_input

See ModifyClusterInput

modify_instance_fleet_input

See ModifyInstanceFleetInput

modify_instance_groups_input

See ModifyInstanceGroupsInput

put_auto_scaling_policy_input

See PutAutoScalingPolicyInput

put_auto_termination_policy_input

See PutAutoTerminationPolicyInput

put_block_public_access_configuration_input

See PutBlockPublicAccessConfigurationInput

put_managed_scaling_policy_input

See PutManagedScalingPolicyInput

remove_auto_scaling_policy_input

See RemoveAutoScalingPolicyInput

remove_auto_termination_policy_input

See RemoveAutoTerminationPolicyInput

remove_managed_scaling_policy_input

See RemoveManagedScalingPolicyInput

remove_tags_input

See RemoveTagsInput

run_job_flow_input

See RunJobFlowInput

set_termination_protection_input

See SetTerminationProtectionInput

set_visible_to_all_users_input

See SetVisibleToAllUsersInput

start_notebook_execution_input

See StartNotebookExecutionInput

stop_notebook_execution_input

See StopNotebookExecutionInput

terminate_job_flows_input

See TerminateJobFlowsInput

update_studio_input

See UpdateStudioInput

update_studio_session_mapping_input

See UpdateStudioSessionMappingInput

Structs

AddInstanceFleetInput

AddInstanceGroupsInput

Input to an AddInstanceGroups call.

AddJobFlowStepsInput

The input argument to the AddJobFlowSteps operation.

AddTagsInput

This input identifies an Amazon EMR resource and a list of tags to attach.

CancelStepsInput

The input argument to the CancelSteps operation.

CreateSecurityConfigurationInput

CreateStudioInput

CreateStudioSessionMappingInput

DeleteSecurityConfigurationInput

DeleteStudioInput

DeleteStudioSessionMappingInput

DescribeClusterInput

This input determines which cluster to describe.

DescribeJobFlowsInput

The input for the DescribeJobFlows operation.

DescribeNotebookExecutionInput

DescribeReleaseLabelInput

DescribeSecurityConfigurationInput

DescribeStepInput

This input determines which step to describe.

DescribeStudioInput

GetAutoTerminationPolicyInput

GetBlockPublicAccessConfigurationInput

GetManagedScalingPolicyInput

GetStudioSessionMappingInput

ListBootstrapActionsInput

This input determines which bootstrap actions to retrieve.

ListClustersInput

This input determines how the ListClusters action filters the list of clusters that it

ListInstanceFleetsInput

ListInstanceGroupsInput

This input determines which instance groups to retrieve.

ListInstancesInput

This input determines which instances to list.

ListNotebookExecutionsInput

ListReleaseLabelsInput

ListSecurityConfigurationsInput

ListStepsInput

This input determines which steps to list.

ListStudioSessionMappingsInput

ListStudiosInput

ModifyClusterInput

ModifyInstanceFleetInput

ModifyInstanceGroupsInput

Change the size of some instance groups.

PutAutoScalingPolicyInput

PutAutoTerminationPolicyInput

PutBlockPublicAccessConfigurationInput

PutManagedScalingPolicyInput

RemoveAutoScalingPolicyInput

RemoveAutoTerminationPolicyInput

RemoveManagedScalingPolicyInput

RemoveTagsInput

This input identifies an Amazon EMR resource and a list of tags to remove.

RunJobFlowInput

Input to the RunJobFlow operation.

SetTerminationProtectionInput

The input argument to the TerminationProtection operation.

SetVisibleToAllUsersInput

The input to the SetVisibleToAllUsers action.

StartNotebookExecutionInput

StopNotebookExecutionInput

TerminateJobFlowsInput

Input to the TerminateJobFlows operation.

UpdateStudioInput

UpdateStudioSessionMappingInput