Struct aws_sdk_emr::input::run_job_flow_input::Builder

#[non_exhaustive]
pub struct Builder { /* fields omitted */ }

A builder for RunJobFlowInput

Implementations

impl Builder

pub fn name(self, input: impl Into<String>) -> Self

The name of the job flow.

pub fn set_name(self, input: Option<String>) -> Self

pub fn log_uri(self, input: impl Into<String>) -> Self

The location in Amazon S3 to write the log files of the job flow. If a value is not provided, logs are not created.

pub fn set_log_uri(self, input: Option<String>) -> Self

pub fn log_encryption_kms_key_id(self, input: impl Into<String>) -> Self

The KMS key used for encrypting log files. If a value is not provided, the logs remain encrypted by AES-256. This attribute is only available with Amazon EMR version 5.30.0 and later, excluding Amazon EMR 6.0.0.

pub fn set_log_encryption_kms_key_id(self, input: Option<String>) -> Self

pub fn additional_info(self, input: impl Into<String>) -> Self

A JSON string for selecting additional features.

pub fn set_additional_info(self, input: Option<String>) -> Self

pub fn ami_version(self, input: impl Into<String>) -> Self

Applies only to Amazon EMR AMI versions 3.x and 2.x. For Amazon EMR releases 4.0 and later, ReleaseLabel is used. To specify a custom AMI, use CustomAmiID.

pub fn set_ami_version(self, input: Option<String>) -> Self

pub fn release_label(self, input: impl Into<String>) -> Self

The Amazon EMR release label, which determines the version of open-source application packages installed on the cluster. Release labels are in the form emr-x.x.x, where x.x.x is an Amazon EMR release version such as emr-5.14.0. For more information about Amazon EMR release versions and included application versions and features, see https://docs.aws.amazon.com/emr/latest/ReleaseGuide/. The release label applies only to Amazon EMR releases version 4.0 and later. Earlier versions use AmiVersion.

pub fn set_release_label(self, input: Option<String>) -> Self

pub fn instances(self, input: JobFlowInstancesConfig) -> Self

A specification of the number and type of Amazon EC2 instances.

pub fn set_instances(self, input: Option<JobFlowInstancesConfig>) -> Self

pub fn steps(self, input: impl Into<StepConfig>) -> Self

pub fn set_steps(self, input: Option<Vec<StepConfig>>) -> Self

pub fn bootstrap_actions(self, input: impl Into<BootstrapActionConfig>) -> Self

pub fn set_bootstrap_actions(
self,
input: Option<Vec<BootstrapActionConfig>>
) -> Self

pub fn supported_products(self, input: impl Into<String>) -> Self

pub fn set_supported_products(self, input: Option<Vec<String>>) -> Self

pub fn new_supported_products(
self,
input: impl Into<SupportedProductConfig>
) -> Self

pub fn set_new_supported_products(
self,
input: Option<Vec<SupportedProductConfig>>
) -> Self

pub fn applications(self, input: impl Into<Application>) -> Self

pub fn set_applications(self, input: Option<Vec<Application>>) -> Self

pub fn configurations(self, input: impl Into<Configuration>) -> Self

pub fn set_configurations(self, input: Option<Vec<Configuration>>) -> Self

pub fn visible_to_all_users(self, input: bool) -> Self

Set this value to true so that IAM principals in the Amazon Web Services account associated with the cluster can perform EMR actions on the cluster that their IAM policies allow. This value defaults to true for clusters created using the EMR API or the CLI create-cluster command.

When set to false, only the IAM principal that created the cluster and the Amazon Web Services account root user can perform EMR actions for the cluster, regardless of the IAM permissions policies attached to other IAM principals. For more information, see Understanding the EMR Cluster VisibleToAllUsers Setting in the Amazon EMRManagement Guide.

pub fn set_visible_to_all_users(self, input: Option<bool>) -> Self

pub fn job_flow_role(self, input: impl Into<String>) -> Self

Also called instance profile and EC2 role. An IAM role for an EMR cluster. The EC2 instances of the cluster assume this role. The default role is EMR_EC2_DefaultRole. In order to use the default role, you must have already created it using the CLI or console.

pub fn set_job_flow_role(self, input: Option<String>) -> Self

pub fn service_role(self, input: impl Into<String>) -> Self

The IAM role that Amazon EMR assumes in order to access Amazon Web Services resources on your behalf.

pub fn set_service_role(self, input: Option<String>) -> Self

pub fn tags(self, input: impl Into<Tag>) -> Self

pub fn set_tags(self, input: Option<Vec<Tag>>) -> Self

pub fn security_configuration(self, input: impl Into<String>) -> Self

The name of a security configuration to apply to the cluster.

pub fn set_security_configuration(self, input: Option<String>) -> Self

pub fn auto_scaling_role(self, input: impl Into<String>) -> Self

An IAM role for automatic scaling policies. The default role is EMR_AutoScaling_DefaultRole. The IAM role provides permissions that the automatic scaling feature requires to launch and terminate EC2 instances in an instance group.

pub fn set_auto_scaling_role(self, input: Option<String>) -> Self

pub fn scale_down_behavior(self, input: ScaleDownBehavior) -> Self

Specifies the way that individual Amazon EC2 instances terminate when an automatic scale-in activity occurs or an instance group is resized. TERMINATE_AT_INSTANCE_HOUR indicates that Amazon EMR terminates nodes at the instance-hour boundary, regardless of when the request to terminate the instance was submitted. This option is only available with Amazon EMR 5.1.0 and later and is the default for clusters created using that version. TERMINATE_AT_TASK_COMPLETION indicates that Amazon EMR adds nodes to a deny list and drains tasks from nodes before terminating the Amazon EC2 instances, regardless of the instance-hour boundary. With either behavior, Amazon EMR removes the least active nodes first and blocks instance termination if it could lead to HDFS corruption. TERMINATE_AT_TASK_COMPLETION available only in Amazon EMR version 4.1.0 and later, and is the default for versions of Amazon EMR earlier than 5.1.0.

pub fn set_scale_down_behavior(self, input: Option<ScaleDownBehavior>) -> Self

pub fn custom_ami_id(self, input: impl Into<String>) -> Self

Available only in Amazon EMR version 5.7.0 and later. The ID of a custom Amazon EBS-backed Linux AMI. If specified, Amazon EMR uses this AMI when it launches cluster EC2 instances. For more information about custom AMIs in Amazon EMR, see Using a Custom AMI in the Amazon EMR Management Guide. If omitted, the cluster uses the base Linux AMI for the ReleaseLabel specified. For Amazon EMR versions 2.x and 3.x, use AmiVersion instead.

For information about creating a custom AMI, see Creating an Amazon EBS-Backed Linux AMI in the Amazon Elastic Compute Cloud User Guide for Linux Instances. For information about finding an AMI ID, see Finding a Linux AMI.

pub fn set_custom_ami_id(self, input: Option<String>) -> Self

pub fn ebs_root_volume_size(self, input: i32) -> Self

The size, in GiB, of the Amazon EBS root device volume of the Linux AMI that is used for each EC2 instance. Available in Amazon EMR version 4.x and later.

pub fn set_ebs_root_volume_size(self, input: Option<i32>) -> Self

pub fn repo_upgrade_on_boot(self, input: RepoUpgradeOnBoot) -> Self

Applies only when CustomAmiID is used. Specifies which updates from the Amazon Linux AMI package repositories to apply automatically when the instance boots using the AMI. If omitted, the default is SECURITY, which indicates that only security updates are applied. If NONE is specified, no updates are applied, and all updates must be applied manually.

pub fn set_repo_upgrade_on_boot(self, input: Option<RepoUpgradeOnBoot>) -> Self

pub fn kerberos_attributes(self, input: KerberosAttributes) -> Self

Attributes for Kerberos configuration when Kerberos authentication is enabled using a security configuration. For more information see Use Kerberos Authentication in the Amazon EMR Management Guide.

pub fn set_kerberos_attributes(self, input: Option<KerberosAttributes>) -> Self

pub fn step_concurrency_level(self, input: i32) -> Self

Specifies the number of steps that can be executed concurrently. The default value is 1. The maximum value is 256.

pub fn set_step_concurrency_level(self, input: Option<i32>) -> Self

pub fn managed_scaling_policy(self, input: ManagedScalingPolicy) -> Self

The specified managed scaling policy for an Amazon EMR cluster.

pub fn set_managed_scaling_policy(
self,
input: Option<ManagedScalingPolicy>
) -> Self

pub fn placement_group_configs(
self,
input: impl Into<PlacementGroupConfig>
) -> Self

pub fn set_placement_group_configs(
self,
input: Option<Vec<PlacementGroupConfig>>
) -> Self

pub fn auto_termination_policy(self, input: AutoTerminationPolicy) -> Self

An auto-termination policy for an Amazon EMR cluster. An auto-termination policy defines the amount of idle time in seconds after which a cluster automatically terminates. For alternative cluster termination options, see Control cluster termination.

pub fn set_auto_termination_policy(
self,
input: Option<AutoTerminationPolicy>
) -> Self

pub fn build(self) -> Result<RunJobFlowInput, BuildError>

Consumes the builder and constructs a RunJobFlowInput