Struct aws_sdk_emr::input::RunJobFlowInput

#[non_exhaustive]
pub struct RunJobFlowInput {

    pub name: Option<String>,
    pub log_uri: Option<String>,
    pub log_encryption_kms_key_id: Option<String>,
    pub additional_info: Option<String>,
    pub ami_version: Option<String>,
    pub release_label: Option<String>,
    pub instances: Option<JobFlowInstancesConfig>,
    pub steps: Option<Vec<StepConfig>>,
    pub bootstrap_actions: Option<Vec<BootstrapActionConfig>>,
    pub supported_products: Option<Vec<String>>,
    pub new_supported_products: Option<Vec<SupportedProductConfig>>,
    pub applications: Option<Vec<Application>>,
    pub configurations: Option<Vec<Configuration>>,
    pub visible_to_all_users: bool,
    pub job_flow_role: Option<String>,
    pub service_role: Option<String>,
    pub tags: Option<Vec<Tag>>,
    pub security_configuration: Option<String>,
    pub auto_scaling_role: Option<String>,
    pub scale_down_behavior: Option<ScaleDownBehavior>,
    pub custom_ami_id: Option<String>,
    pub ebs_root_volume_size: Option<i32>,
    pub repo_upgrade_on_boot: Option<RepoUpgradeOnBoot>,
    pub kerberos_attributes: Option<KerberosAttributes>,
    pub step_concurrency_level: Option<i32>,
    pub managed_scaling_policy: Option<ManagedScalingPolicy>,
    pub placement_group_configs: Option<Vec<PlacementGroupConfig>>,
    pub auto_termination_policy: Option<AutoTerminationPolicy>,

}

Input to the RunJobFlow operation.

Fields (Non-exhaustive)

Non-exhaustive structs could have additional fields added in future. Therefore, non-exhaustive structs cannot be constructed in external crates using the traditional Struct { .. } syntax; cannot be matched against without a wildcard ..; and struct update syntax will not work.

name: Option<String>

The name of the job flow.

log_uri: Option<String>

The location in Amazon S3 to write the log files of the job flow. If a value is not provided, logs are not created.

log_encryption_kms_key_id: Option<String>

The KMS key used for encrypting log files. If a value is not provided, the logs remain encrypted by AES-256. This attribute is only available with Amazon EMR version 5.30.0 and later, excluding Amazon EMR 6.0.0.

additional_info: Option<String>

A JSON string for selecting additional features.

ami_version: Option<String>

Applies only to Amazon EMR AMI versions 3.x and 2.x. For Amazon EMR releases 4.0 and later, ReleaseLabel is used. To specify a custom AMI, use CustomAmiID.

release_label: Option<String>

The Amazon EMR release label, which determines the version of open-source application packages installed on the cluster. Release labels are in the form emr-x.x.x, where x.x.x is an Amazon EMR release version such as emr-5.14.0. For more information about Amazon EMR release versions and included application versions and features, see https://docs.aws.amazon.com/emr/latest/ReleaseGuide/. The release label applies only to Amazon EMR releases version 4.0 and later. Earlier versions use AmiVersion.

instances: Option<JobFlowInstancesConfig>

A specification of the number and type of Amazon EC2 instances.

steps: Option<Vec<StepConfig>>

A list of steps to run.

bootstrap_actions: Option<Vec<BootstrapActionConfig>>

A list of bootstrap actions to run before Hadoop starts on the cluster nodes.

supported_products: Option<Vec<String>>

For Amazon EMR releases 3.x and 2.x. For Amazon EMR releases 4.x and later, use Applications.

A list of strings that indicates third-party software to use. For more information, see the Amazon EMR Developer Guide. Currently supported values are:

"mapr-m3" - launch the job flow using MapR M3 Edition.

"mapr-m5" - launch the job flow using MapR M5 Edition.

new_supported_products: Option<Vec<SupportedProductConfig>>

For Amazon EMR releases 3.x and 2.x. For Amazon EMR releases 4.x and later, use Applications.

A list of strings that indicates third-party software to use with the job flow that accepts a user argument list. EMR accepts and forwards the argument list to the corresponding installation script as bootstrap action arguments. For more information, see "Launch a Job Flow on the MapR Distribution for Hadoop" in the Amazon EMR Developer Guide. Supported values are:

"mapr-m3" - launch the cluster using MapR M3 Edition.

"mapr-m5" - launch the cluster using MapR M5 Edition.

"mapr" with the user arguments specifying "--edition,m3" or "--edition,m5" - launch the job flow using MapR M3 or M5 Edition respectively.

"mapr-m7" - launch the cluster using MapR M7 Edition.

"hunk" - launch the cluster with the Hunk Big Data Analytics Platform.

"hue"- launch the cluster with Hue installed.

"spark" - launch the cluster with Apache Spark installed.

"ganglia" - launch the cluster with the Ganglia Monitoring System installed.

applications: Option<Vec<Application>>

Applies to Amazon EMR releases 4.0 and later. A case-insensitive list of applications for Amazon EMR to install and configure when launching the cluster. For a list of applications available for each Amazon EMR release version, see the Amazon EMR Release Guide.

configurations: Option<Vec<Configuration>>

For Amazon EMR releases 4.0 and later. The list of configurations supplied for the EMR cluster you are creating.

visible_to_all_users: bool

Set this value to true so that IAM principals in the Amazon Web Services account associated with the cluster can perform EMR actions on the cluster that their IAM policies allow. This value defaults to true for clusters created using the EMR API or the CLI create-cluster command.

When set to false, only the IAM principal that created the cluster and the Amazon Web Services account root user can perform EMR actions for the cluster, regardless of the IAM permissions policies attached to other IAM principals. For more information, see Understanding the EMR Cluster VisibleToAllUsers Setting in the Amazon EMRManagement Guide.

job_flow_role: Option<String>

Also called instance profile and EC2 role. An IAM role for an EMR cluster. The EC2 instances of the cluster assume this role. The default role is EMR_EC2_DefaultRole. In order to use the default role, you must have already created it using the CLI or console.

service_role: Option<String>

The IAM role that Amazon EMR assumes in order to access Amazon Web Services resources on your behalf.

tags: Option<Vec<Tag>>

A list of tags to associate with a cluster and propagate to Amazon EC2 instances.

security_configuration: Option<String>

The name of a security configuration to apply to the cluster.

auto_scaling_role: Option<String>

An IAM role for automatic scaling policies. The default role is EMR_AutoScaling_DefaultRole. The IAM role provides permissions that the automatic scaling feature requires to launch and terminate EC2 instances in an instance group.

scale_down_behavior: Option<ScaleDownBehavior>

Specifies the way that individual Amazon EC2 instances terminate when an automatic scale-in activity occurs or an instance group is resized. TERMINATE_AT_INSTANCE_HOUR indicates that Amazon EMR terminates nodes at the instance-hour boundary, regardless of when the request to terminate the instance was submitted. This option is only available with Amazon EMR 5.1.0 and later and is the default for clusters created using that version. TERMINATE_AT_TASK_COMPLETION indicates that Amazon EMR adds nodes to a deny list and drains tasks from nodes before terminating the Amazon EC2 instances, regardless of the instance-hour boundary. With either behavior, Amazon EMR removes the least active nodes first and blocks instance termination if it could lead to HDFS corruption. TERMINATE_AT_TASK_COMPLETION available only in Amazon EMR version 4.1.0 and later, and is the default for versions of Amazon EMR earlier than 5.1.0.

custom_ami_id: Option<String>

Available only in Amazon EMR version 5.7.0 and later. The ID of a custom Amazon EBS-backed Linux AMI. If specified, Amazon EMR uses this AMI when it launches cluster EC2 instances. For more information about custom AMIs in Amazon EMR, see Using a Custom AMI in the Amazon EMR Management Guide. If omitted, the cluster uses the base Linux AMI for the ReleaseLabel specified. For Amazon EMR versions 2.x and 3.x, use AmiVersion instead.

For information about creating a custom AMI, see Creating an Amazon EBS-Backed Linux AMI in the Amazon Elastic Compute Cloud User Guide for Linux Instances. For information about finding an AMI ID, see Finding a Linux AMI.

ebs_root_volume_size: Option<i32>

The size, in GiB, of the Amazon EBS root device volume of the Linux AMI that is used for each EC2 instance. Available in Amazon EMR version 4.x and later.

repo_upgrade_on_boot: Option<RepoUpgradeOnBoot>

Applies only when CustomAmiID is used. Specifies which updates from the Amazon Linux AMI package repositories to apply automatically when the instance boots using the AMI. If omitted, the default is SECURITY, which indicates that only security updates are applied. If NONE is specified, no updates are applied, and all updates must be applied manually.

kerberos_attributes: Option<KerberosAttributes>

Attributes for Kerberos configuration when Kerberos authentication is enabled using a security configuration. For more information see Use Kerberos Authentication in the Amazon EMR Management Guide.

step_concurrency_level: Option<i32>

Specifies the number of steps that can be executed concurrently. The default value is 1. The maximum value is 256.

managed_scaling_policy: Option<ManagedScalingPolicy>

The specified managed scaling policy for an Amazon EMR cluster.

placement_group_configs: Option<Vec<PlacementGroupConfig>>

The specified placement group configuration for an Amazon EMR cluster.

auto_termination_policy: Option<AutoTerminationPolicy>

An auto-termination policy for an Amazon EMR cluster. An auto-termination policy defines the amount of idle time in seconds after which a cluster automatically terminates. For alternative cluster termination options, see Control cluster termination.

Implementations

impl RunJobFlowInput

pub fn make_operation(
&self,
_config: &Config
) -> Result<Operation<RunJobFlow, AwsErrorRetryPolicy>, BuildError>

Consumes the builder and constructs an Operation<RunJobFlow>

pub fn builder() -> Builder

Creates a new builder-style object to manufacture RunJobFlowInput