Module aws_sdk_emr::model

Modules

application

See Application

auto_scaling_policy

See AutoScalingPolicy

auto_scaling_policy_description

See AutoScalingPolicyDescription

auto_scaling_policy_state_change_reason

See AutoScalingPolicyStateChangeReason

auto_scaling_policy_status

See AutoScalingPolicyStatus

auto_termination_policy

See AutoTerminationPolicy

block_public_access_configuration

See BlockPublicAccessConfiguration

block_public_access_configuration_metadata

See BlockPublicAccessConfigurationMetadata

bootstrap_action_config

See BootstrapActionConfig

bootstrap_action_detail

See BootstrapActionDetail

cancel_steps_info

See CancelStepsInfo

cloud_watch_alarm_definition

See CloudWatchAlarmDefinition

cluster

See Cluster

cluster_state_change_reason

See ClusterStateChangeReason

cluster_status

See ClusterStatus

cluster_summary

See ClusterSummary

cluster_timeline

See ClusterTimeline

command

See Command

compute_limits

See ComputeLimits

configuration

See Configuration

ebs_block_device

See EbsBlockDevice

ebs_block_device_config

See EbsBlockDeviceConfig

ebs_configuration

See EbsConfiguration

ebs_volume

See EbsVolume

ec2_instance_attributes

See Ec2InstanceAttributes

execution_engine_config

See ExecutionEngineConfig

failure_details

See FailureDetails

hadoop_jar_step_config

See HadoopJarStepConfig

hadoop_step_config

See HadoopStepConfig

instance

See Instance

instance_fleet

See InstanceFleet

instance_fleet_config

See InstanceFleetConfig

instance_fleet_modify_config

See InstanceFleetModifyConfig

instance_fleet_provisioning_specifications

See InstanceFleetProvisioningSpecifications

instance_fleet_state_change_reason

See InstanceFleetStateChangeReason

instance_fleet_status

See InstanceFleetStatus

instance_fleet_timeline

See InstanceFleetTimeline

instance_group

See InstanceGroup

instance_group_config

See InstanceGroupConfig

instance_group_detail

See InstanceGroupDetail

instance_group_modify_config

See InstanceGroupModifyConfig

instance_group_state_change_reason

See InstanceGroupStateChangeReason

instance_group_status

See InstanceGroupStatus

instance_group_timeline

See InstanceGroupTimeline

instance_resize_policy

See InstanceResizePolicy

instance_state_change_reason

See InstanceStateChangeReason

instance_status

See InstanceStatus

instance_timeline

See InstanceTimeline

instance_type_config

See InstanceTypeConfig

instance_type_specification

See InstanceTypeSpecification

job_flow_detail

See JobFlowDetail

job_flow_execution_status_detail

See JobFlowExecutionStatusDetail

job_flow_instances_config

See JobFlowInstancesConfig

job_flow_instances_detail

See JobFlowInstancesDetail

kerberos_attributes

See KerberosAttributes

key_value

See KeyValue

managed_scaling_policy

See ManagedScalingPolicy

metric_dimension

See MetricDimension

notebook_execution

See NotebookExecution

notebook_execution_summary

See NotebookExecutionSummary

on_demand_capacity_reservation_options

See OnDemandCapacityReservationOptions

on_demand_provisioning_specification

See OnDemandProvisioningSpecification

placement_group_config

See PlacementGroupConfig

placement_type

See PlacementType

port_range

See PortRange

release_label_filter

See ReleaseLabelFilter

scaling_action

See ScalingAction

scaling_constraints

See ScalingConstraints

scaling_rule

See ScalingRule

scaling_trigger

See ScalingTrigger

script_bootstrap_action_config

See ScriptBootstrapActionConfig

security_configuration_summary

See SecurityConfigurationSummary

session_mapping_detail

See SessionMappingDetail

session_mapping_summary

See SessionMappingSummary

shrink_policy

See ShrinkPolicy

simple_scaling_policy_configuration

See SimpleScalingPolicyConfiguration

simplified_application

See SimplifiedApplication

spot_provisioning_specification

See SpotProvisioningSpecification

step

See Step

step_config

See StepConfig

step_detail

See StepDetail

step_execution_status_detail

See StepExecutionStatusDetail

step_state_change_reason

See StepStateChangeReason

step_status

See StepStatus

step_summary

See StepSummary

step_timeline

See StepTimeline

studio

See Studio

studio_summary

See StudioSummary

supported_product_config

See SupportedProductConfig

tag

See Tag

volume_specification

See VolumeSpecification

Structs

Application

With Amazon EMR release version 4.0 and later, the only accepted parameter is the

AutoScalingPolicy

An automatic scaling policy for a core instance group or task instance group in an

AutoScalingPolicyDescription

An automatic scaling policy for a core instance group or task instance group in an

AutoScalingPolicyStateChangeReason

The reason for an AutoScalingPolicyStatus change.

AutoScalingPolicyStatus

The status of an automatic scaling policy.

AutoTerminationPolicy

An auto-termination policy for an Amazon EMR cluster. An auto-termination policy defines the amount of idle time in seconds after which a cluster automatically terminates. For alternative cluster termination options, see Control cluster termination.

BlockPublicAccessConfiguration

A configuration for Amazon EMR block public access. When

BlockPublicAccessConfigurationMetadata

Properties that describe the Amazon Web Services principal that created the

BootstrapActionConfig

Configuration of a bootstrap action.

BootstrapActionDetail

Reports the configuration of a bootstrap action in a cluster (job flow).

CancelStepsInfo

Specification of the status of a CancelSteps request. Available only in Amazon EMR

CloudWatchAlarmDefinition

The definition of a CloudWatch metric alarm, which determines when an automatic scaling

Cluster

The detailed description of the cluster.

ClusterStateChangeReason

The reason that the cluster changed to its current state.

ClusterStatus

The detailed status of the cluster.

ClusterSummary

The summary description of the cluster.

ClusterTimeline

Represents the timeline of the cluster's lifecycle.

Command

An entity describing an executable that runs on a cluster.

ComputeLimits

The EC2 unit limits for a managed scaling policy. The managed scaling activity of a

Configuration

EbsBlockDevice

Configuration of requested EBS block device associated with the instance group.

EbsBlockDeviceConfig

Configuration of requested EBS block device associated with the instance group with

EbsConfiguration

The Amazon EBS configuration of a cluster instance.

EbsVolume

EBS block device that's attached to an EC2 instance.

Ec2InstanceAttributes

Provides information about the EC2 instances in a cluster grouped by category. For

ExecutionEngineConfig

Specifies the execution engine (cluster) to run the notebook and perform the notebook

FailureDetails

The details of the step failure. The service attempts to detect the root cause for many

HadoopJarStepConfig

A job flow step consisting of a JAR file whose main function will be executed. The main

HadoopStepConfig

A cluster step consisting of a JAR file whose main function will be executed. The main

Instance

Represents an EC2 instance provisioned as part of cluster.

InstanceFleet

Describes an instance fleet, which is a group of EC2 instances that host a particular

InstanceFleetConfig

The configuration that defines an instance fleet.

InstanceFleetModifyConfig

Configuration parameters for an instance fleet modification request.

InstanceFleetProvisioningSpecifications

The launch specification for Spot Instances in the fleet, which determines the defined

InstanceFleetStateChangeReason

Provides status change reason details for the instance fleet.

InstanceFleetStatus

The status of the instance fleet.

InstanceFleetTimeline

Provides historical timestamps for the instance fleet, including the time of creation,

InstanceGroup

This entity represents an instance group, which is a group of instances that have common

InstanceGroupConfig

Configuration defining a new instance group.

InstanceGroupDetail

Detailed information about an instance group.

InstanceGroupModifyConfig

Modify the size or configurations of an instance group.

InstanceGroupStateChangeReason

The status change reason details for the instance group.

InstanceGroupStatus

The details of the instance group status.

InstanceGroupTimeline

The timeline of the instance group lifecycle.

InstanceResizePolicy

Custom policy for requesting termination protection or termination of specific instances

InstanceStateChangeReason

The details of the status change reason for the instance.

InstanceStatus

The instance status details.

InstanceTimeline

The timeline of the instance lifecycle.

InstanceTypeConfig

An instance type configuration for each instance type in an instance fleet, which

InstanceTypeSpecification

The configuration specification for each instance type in an instance fleet.

JobFlowDetail

A description of a cluster (job flow).

JobFlowExecutionStatusDetail

Describes the status of the cluster (job flow).

JobFlowInstancesConfig

A description of the Amazon EC2 instance on which the cluster (job flow) runs. A valid

JobFlowInstancesDetail

Specify the type of Amazon EC2 instances that the cluster (job flow) runs on.

KerberosAttributes

Attributes for Kerberos configuration when Kerberos authentication is enabled using a

KeyValue

A key-value pair.

ManagedScalingPolicy

Managed scaling policy for an Amazon EMR cluster. The policy specifies the limits for

MetricDimension

A CloudWatch dimension, which is specified using a Key (known as a

NotebookExecution

A notebook execution. An execution is a specific instance that an EMR Notebook is run

NotebookExecutionSummary

Details for a notebook execution. The details include information such as the unique ID and status of the notebook execution.

OnDemandCapacityReservationOptions

Describes the strategy for using unused Capacity Reservations for fulfilling On-Demand capacity.

OnDemandProvisioningSpecification

The launch specification for On-Demand Instances in the instance fleet, which

PlacementGroupConfig

Placement group configuration for an Amazon EMR cluster. The configuration specifies the

PlacementType

The Amazon EC2 Availability Zone configuration of the cluster (job flow).

PortRange

A list of port ranges that are permitted to allow inbound traffic from all public IP

ReleaseLabelFilter

The release label filters by application or version prefix.

ScalingAction

The type of adjustment the automatic scaling activity makes when triggered, and the

ScalingConstraints

The upper and lower EC2 instance limits for an automatic scaling policy. Automatic

ScalingRule

A scale-in or scale-out rule that defines scaling activity, including the CloudWatch

ScalingTrigger

The conditions that trigger an automatic scaling activity.

ScriptBootstrapActionConfig

Configuration of the script to run during a bootstrap action.

SecurityConfigurationSummary

The creation date and time, and name, of a security configuration.

SessionMappingDetail

Details for an Amazon EMR Studio session mapping including creation time, user or group

SessionMappingSummary

Details for an Amazon EMR Studio session mapping. The details do not include the time

ShrinkPolicy

Policy for customizing shrink operations. Allows configuration of decommissioning

SimpleScalingPolicyConfiguration

An automatic scaling configuration, which describes how the policy adds or removes

SimplifiedApplication

The returned release label application names or versions.

SpotProvisioningSpecification

The launch specification for Spot Instances in the instance fleet, which determines the

Step

This represents a step in a cluster.

StepConfig

Specification for a cluster (job flow) step.

StepDetail

Combines the execution state and configuration of a step.

StepExecutionStatusDetail

The execution state of a step.

StepStateChangeReason

The details of the step state change reason.

StepStatus

The execution status details of the cluster step.

StepSummary

The summary of the cluster step.

StepTimeline

The timeline of the cluster step lifecycle.

Studio

Details for an Amazon EMR Studio including ID, creation time, name, and so on.

StudioSummary

Details for an Amazon EMR Studio, including ID, Name, VPC, and Description. The details

SupportedProductConfig

The list of supported product configurations that allow user-supplied arguments. EMR

Tag

A key-value pair containing user-defined metadata that you can associate with an Amazon EMR resource. Tags make it easier to associate clusters in various ways, such as grouping

VolumeSpecification

EBS volume specifications such as volume type, IOPS, and size (GiB) that will be

Enums

ActionOnFailure

AdjustmentType

AuthMode

AutoScalingPolicyState

AutoScalingPolicyStateChangeReasonCode

CancelStepsRequestStatus

ClusterState

ClusterStateChangeReasonCode

ComparisonOperator

ComputeLimitsUnitType

ExecutionEngineType

IdentityType

InstanceCollectionType

InstanceFleetState

InstanceFleetStateChangeReasonCode

InstanceFleetType

InstanceGroupState

InstanceGroupStateChangeReasonCode

InstanceGroupType

InstanceRoleType

InstanceState

InstanceStateChangeReasonCode

JobFlowExecutionState

The type of instance.

MarketType

NotebookExecutionStatus

OnDemandCapacityReservationPreference

OnDemandCapacityReservationUsageStrategy

OnDemandProvisioningAllocationStrategy

PlacementGroupStrategy

RepoUpgradeOnBoot

ScaleDownBehavior

SpotProvisioningAllocationStrategy

SpotProvisioningTimeoutAction

Statistic

StepCancellationOption

StepExecutionState

StepState

StepStateChangeReasonCode

Unit