Struct aws_sdk_emr::model::cluster::Builder

#[non_exhaustive]
pub struct Builder { /* fields omitted */ }

A builder for Cluster

Implementations

impl Builder

pub fn id(self, input: impl Into<String>) -> Self

The unique identifier for the cluster.

pub fn set_id(self, input: Option<String>) -> Self

pub fn name(self, input: impl Into<String>) -> Self

The name of the cluster.

pub fn set_name(self, input: Option<String>) -> Self

pub fn status(self, input: ClusterStatus) -> Self

The current status details about the cluster.

pub fn set_status(self, input: Option<ClusterStatus>) -> Self

pub fn ec2_instance_attributes(self, input: Ec2InstanceAttributes) -> Self

Provides information about the EC2 instances in a cluster grouped by category. For example, key name, subnet ID, IAM instance profile, and so on.

pub fn set_ec2_instance_attributes(
self,
input: Option<Ec2InstanceAttributes>
) -> Self

pub fn instance_collection_type(self, input: InstanceCollectionType) -> Self

The instance fleet configuration is available only in Amazon EMR versions 4.8.0 and later, excluding 5.0.x versions.

The instance group configuration of the cluster. A value of INSTANCE_GROUP indicates a uniform instance group configuration. A value of INSTANCE_FLEET indicates an instance fleets configuration.

pub fn set_instance_collection_type(
self,
input: Option<InstanceCollectionType>
) -> Self

pub fn log_uri(self, input: impl Into<String>) -> Self

The path to the Amazon S3 location where logs for this cluster are stored.

pub fn set_log_uri(self, input: Option<String>) -> Self

pub fn log_encryption_kms_key_id(self, input: impl Into<String>) -> Self

The KMS key used for encrypting log files. This attribute is only available with EMR version 5.30.0 and later, excluding EMR 6.0.0.

pub fn set_log_encryption_kms_key_id(self, input: Option<String>) -> Self

pub fn requested_ami_version(self, input: impl Into<String>) -> Self

The AMI version requested for this cluster.

pub fn set_requested_ami_version(self, input: Option<String>) -> Self

pub fn running_ami_version(self, input: impl Into<String>) -> Self

The AMI version running on this cluster.

pub fn set_running_ami_version(self, input: Option<String>) -> Self

pub fn release_label(self, input: impl Into<String>) -> Self

The Amazon EMR release label, which determines the version of open-source application packages installed on the cluster. Release labels are in the form emr-x.x.x, where x.x.x is an Amazon EMR release version such as emr-5.14.0. For more information about Amazon EMR release versions and included application versions and features, see https://docs.aws.amazon.com/emr/latest/ReleaseGuide/. The release label applies only to Amazon EMR releases version 4.0 and later. Earlier versions use AmiVersion.

pub fn set_release_label(self, input: Option<String>) -> Self

pub fn auto_terminate(self, input: bool) -> Self

Specifies whether the cluster should terminate after completing all steps.

pub fn set_auto_terminate(self, input: Option<bool>) -> Self

pub fn termination_protected(self, input: bool) -> Self

Indicates whether Amazon EMR will lock the cluster to prevent the EC2 instances from being terminated by an API call or user intervention, or in the event of a cluster error.

pub fn set_termination_protected(self, input: Option<bool>) -> Self

pub fn visible_to_all_users(self, input: bool) -> Self

Indicates whether the cluster is visible to IAM principals in the Amazon Web Services account associated with the cluster. When true, IAM principals in the Amazon Web Services account can perform EMR cluster actions on the cluster that their IAM policies allow. When false, only the IAM principal that created the cluster and the Amazon Web Services account root user can perform EMR actions, regardless of IAM permissions policies attached to other IAM principals.

The default value is true if a value is not provided when creating a cluster using the EMR API RunJobFlow command, the CLI create-cluster command, or the Amazon Web Services Management Console. IAM principals that are allowed to perform actions on the cluster can use the SetVisibleToAllUsers action to change the value on a running cluster. For more information, see Understanding the EMR Cluster VisibleToAllUsers Setting in the Amazon EMRManagement Guide.

pub fn set_visible_to_all_users(self, input: Option<bool>) -> Self

pub fn applications(self, input: impl Into<Application>) -> Self

pub fn set_applications(self, input: Option<Vec<Application>>) -> Self

pub fn tags(self, input: impl Into<Tag>) -> Self

pub fn set_tags(self, input: Option<Vec<Tag>>) -> Self

pub fn service_role(self, input: impl Into<String>) -> Self

The IAM role that Amazon EMR assumes in order to access Amazon Web Services resources on your behalf.

pub fn set_service_role(self, input: Option<String>) -> Self

pub fn normalized_instance_hours(self, input: i32) -> Self

An approximation of the cost of the cluster, represented in m1.small/hours. This value is incremented one time for every hour an m1.small instance runs. Larger instances are weighted more, so an EC2 instance that is roughly four times more expensive would result in the normalized instance hours being incremented by four. This result is only an approximation and does not reflect the actual billing rate.

pub fn set_normalized_instance_hours(self, input: Option<i32>) -> Self

pub fn master_public_dns_name(self, input: impl Into<String>) -> Self

The DNS name of the master node. If the cluster is on a private subnet, this is the private DNS name. On a public subnet, this is the public DNS name.

pub fn set_master_public_dns_name(self, input: Option<String>) -> Self

pub fn configurations(self, input: impl Into<Configuration>) -> Self

pub fn set_configurations(self, input: Option<Vec<Configuration>>) -> Self

pub fn security_configuration(self, input: impl Into<String>) -> Self

The name of the security configuration applied to the cluster.

pub fn set_security_configuration(self, input: Option<String>) -> Self

pub fn auto_scaling_role(self, input: impl Into<String>) -> Self

An IAM role for automatic scaling policies. The default role is EMR_AutoScaling_DefaultRole. The IAM role provides permissions that the automatic scaling feature requires to launch and terminate EC2 instances in an instance group.

pub fn set_auto_scaling_role(self, input: Option<String>) -> Self

pub fn scale_down_behavior(self, input: ScaleDownBehavior) -> Self

The way that individual Amazon EC2 instances terminate when an automatic scale-in activity occurs or an instance group is resized. TERMINATE_AT_INSTANCE_HOUR indicates that Amazon EMR terminates nodes at the instance-hour boundary, regardless of when the request to terminate the instance was submitted. This option is only available with Amazon EMR 5.1.0 and later and is the default for clusters created using that version. TERMINATE_AT_TASK_COMPLETION indicates that Amazon EMR adds nodes to a deny list and drains tasks from nodes before terminating the Amazon EC2 instances, regardless of the instance-hour boundary. With either behavior, Amazon EMR removes the least active nodes first and blocks instance termination if it could lead to HDFS corruption. TERMINATE_AT_TASK_COMPLETION is available only in Amazon EMR version 4.1.0 and later, and is the default for versions of Amazon EMR earlier than 5.1.0.

pub fn set_scale_down_behavior(self, input: Option<ScaleDownBehavior>) -> Self

pub fn custom_ami_id(self, input: impl Into<String>) -> Self

Available only in Amazon EMR version 5.7.0 and later. The ID of a custom Amazon EBS-backed Linux AMI if the cluster uses a custom AMI.

pub fn set_custom_ami_id(self, input: Option<String>) -> Self

pub fn ebs_root_volume_size(self, input: i32) -> Self

The size, in GiB, of the Amazon EBS root device volume of the Linux AMI that is used for each EC2 instance. Available in Amazon EMR version 4.x and later.

pub fn set_ebs_root_volume_size(self, input: Option<i32>) -> Self

pub fn repo_upgrade_on_boot(self, input: RepoUpgradeOnBoot) -> Self

Applies only when CustomAmiID is used. Specifies the type of updates that are applied from the Amazon Linux AMI package repositories when an instance boots using the AMI.

pub fn set_repo_upgrade_on_boot(self, input: Option<RepoUpgradeOnBoot>) -> Self

pub fn kerberos_attributes(self, input: KerberosAttributes) -> Self

Attributes for Kerberos configuration when Kerberos authentication is enabled using a security configuration. For more information see Use Kerberos Authentication in the Amazon EMR Management Guide.

pub fn set_kerberos_attributes(self, input: Option<KerberosAttributes>) -> Self

pub fn cluster_arn(self, input: impl Into<String>) -> Self

The Amazon Resource Name of the cluster.

pub fn set_cluster_arn(self, input: Option<String>) -> Self

pub fn outpost_arn(self, input: impl Into<String>) -> Self

The Amazon Resource Name (ARN) of the Outpost where the cluster is launched.

pub fn set_outpost_arn(self, input: Option<String>) -> Self

pub fn step_concurrency_level(self, input: i32) -> Self

Specifies the number of steps that can be executed concurrently.

pub fn set_step_concurrency_level(self, input: Option<i32>) -> Self

pub fn placement_groups(self, input: impl Into<PlacementGroupConfig>) -> Self

pub fn set_placement_groups(
self,
input: Option<Vec<PlacementGroupConfig>>
) -> Self

pub fn build(self) -> Cluster

Consumes the builder and constructs a Cluster