Module aws_sdk_emr::operation

Structs

AddInstanceFleet

Adds an instance fleet to a running cluster.

AddInstanceGroups

Adds one or more instance groups to a running cluster.

AddJobFlowSteps

AddJobFlowSteps adds new steps to a running cluster. A maximum of 256 steps are allowed

AddTags

Adds tags to an Amazon EMR resource, such as a cluster or an Amazon EMR Studio. Tags make it easier to associate resources in

CancelSteps

Cancels a pending step or steps in a running cluster. Available only in Amazon EMR

CreateSecurityConfiguration

Creates a security configuration, which is stored in the service and can be specified

CreateStudio

Creates a new Amazon EMR Studio.

CreateStudioSessionMapping

Maps a user or group to the Amazon EMR Studio specified by StudioId, and

DeleteSecurityConfiguration

Deletes a security configuration.

DeleteStudio

Removes an Amazon EMR Studio from the Studio metadata store.

DeleteStudioSessionMapping

Removes a user or group from an Amazon EMR Studio.

DescribeCluster

Provides cluster-level details including status, hardware and software configuration,

DescribeJobFlows

This API is no longer supported and will eventually be removed. We recommend you use

DescribeNotebookExecution

Provides details of a notebook execution.

DescribeReleaseLabel

Provides EMR release label details, such as releases available the region where the API request is run, and the available applications for a specific EMR release label. Can also list EMR release versions that support a specified version of Spark.

DescribeSecurityConfiguration

Provides the details of a security configuration by returning the configuration

DescribeStep

Provides more detail about the cluster step.

DescribeStudio

Returns details for the specified Amazon EMR Studio including ID, Name, VPC, Studio

GetAutoTerminationPolicy

Returns the auto-termination policy for an Amazon EMR cluster.

GetBlockPublicAccessConfiguration

Returns the Amazon EMR block public access configuration for your Amazon Web Services account in the

GetManagedScalingPolicy

Fetches the attached managed scaling policy for an Amazon EMR cluster.

GetStudioSessionMapping

Fetches mapping details for the specified Amazon EMR Studio and identity (user or

ListBootstrapActions

Provides information about the bootstrap actions associated with a cluster.

ListClusters

Provides the status of all clusters visible to this Amazon Web Services account. Allows you to filter

ListInstanceFleets

Lists all available details about the instance fleets in a cluster.

ListInstanceGroups

Provides all available details about the instance groups in a cluster.

ListInstances

Provides information for all active EC2 instances and EC2 instances terminated in the

ListNotebookExecutions

Provides summaries of all notebook executions. You can filter the list based on multiple

ListReleaseLabels

Retrieves release labels of EMR services in the region where the API is called.

ListSecurityConfigurations

Lists all the security configurations visible to this account, providing their creation

ListSteps

Provides a list of steps for the cluster in reverse order unless you specify

ListStudioSessionMappings

Returns a list of all user or group session mappings for the Amazon EMR Studio specified by

ListStudios

Returns a list of all Amazon EMR Studios associated with the Amazon Web Services account. The list

ModifyCluster

Modifies the number of steps that can be executed concurrently for the cluster specified

ModifyInstanceFleet

Modifies the target On-Demand and target Spot capacities for the instance fleet with the

ModifyInstanceGroups

ModifyInstanceGroups modifies the number of nodes and configuration settings of an

PutAutoScalingPolicy

Creates or updates an automatic scaling policy for a core instance group or task

PutAutoTerminationPolicy

Creates or updates an auto-termination policy for an Amazon EMR cluster. An auto-termination policy defines the amount of idle time in seconds after which a cluster automatically terminates. For alternative cluster termination options, see Control cluster termination.

PutBlockPublicAccessConfiguration

Creates or updates an Amazon EMR block public access configuration for your Amazon Web Services account

PutManagedScalingPolicy

Creates or updates a managed scaling policy for an Amazon EMR cluster. The managed

RemoveAutoScalingPolicy

Removes an automatic scaling policy from a specified instance group within an EMR

RemoveAutoTerminationPolicy

Removes an auto-termination policy from an Amazon EMR cluster.

RemoveManagedScalingPolicy

Removes a managed scaling policy from a specified EMR cluster.

RemoveTags

Removes tags from an Amazon EMR resource, such as a cluster or Amazon EMR Studio. Tags make it easier to associate resources in

RunJobFlow

RunJobFlow creates and starts running a new cluster (job flow). The cluster runs the

SetTerminationProtection

SetTerminationProtection locks a cluster (job flow) so the EC2 instances in the cluster

SetVisibleToAllUsers

Sets the Cluster$VisibleToAllUsers value for an EMR cluster. When true, IAM principals in the

StartNotebookExecution

Starts a notebook execution.

StopNotebookExecution

Stops a notebook execution.

TerminateJobFlows

TerminateJobFlows shuts a list of clusters (job flows) down. When a job flow is shut

UpdateStudio

Updates an Amazon EMR Studio configuration, including attributes such as name, description, and subnets.

UpdateStudioSessionMapping

Updates the session policy attached to the user or group for the specified Amazon EMR