Module aws_sdk_emr::output

Modules

add_instance_fleet_output

See AddInstanceFleetOutput

add_instance_groups_output

See AddInstanceGroupsOutput

add_job_flow_steps_output

See AddJobFlowStepsOutput

add_tags_output

See AddTagsOutput

cancel_steps_output

See CancelStepsOutput

create_security_configuration_output

See CreateSecurityConfigurationOutput

create_studio_output

See CreateStudioOutput

create_studio_session_mapping_output

See CreateStudioSessionMappingOutput

delete_security_configuration_output

See DeleteSecurityConfigurationOutput

delete_studio_output

See DeleteStudioOutput

delete_studio_session_mapping_output

See DeleteStudioSessionMappingOutput

describe_cluster_output

See DescribeClusterOutput

describe_job_flows_output

See DescribeJobFlowsOutput

describe_notebook_execution_output

See DescribeNotebookExecutionOutput

describe_release_label_output

See DescribeReleaseLabelOutput

describe_security_configuration_output

See DescribeSecurityConfigurationOutput

describe_step_output

See DescribeStepOutput

describe_studio_output

See DescribeStudioOutput

get_auto_termination_policy_output

See GetAutoTerminationPolicyOutput

get_block_public_access_configuration_output

See GetBlockPublicAccessConfigurationOutput

get_managed_scaling_policy_output

See GetManagedScalingPolicyOutput

get_studio_session_mapping_output

See GetStudioSessionMappingOutput

list_bootstrap_actions_output

See ListBootstrapActionsOutput

list_clusters_output

See ListClustersOutput

list_instance_fleets_output

See ListInstanceFleetsOutput

list_instance_groups_output

See ListInstanceGroupsOutput

list_instances_output

See ListInstancesOutput

list_notebook_executions_output

See ListNotebookExecutionsOutput

list_release_labels_output

See ListReleaseLabelsOutput

list_security_configurations_output

See ListSecurityConfigurationsOutput

list_steps_output

See ListStepsOutput

list_studio_session_mappings_output

See ListStudioSessionMappingsOutput

list_studios_output

See ListStudiosOutput

modify_cluster_output

See ModifyClusterOutput

modify_instance_fleet_output

See ModifyInstanceFleetOutput

modify_instance_groups_output

See ModifyInstanceGroupsOutput

put_auto_scaling_policy_output

See PutAutoScalingPolicyOutput

put_auto_termination_policy_output

See PutAutoTerminationPolicyOutput

put_block_public_access_configuration_output

See PutBlockPublicAccessConfigurationOutput

put_managed_scaling_policy_output

See PutManagedScalingPolicyOutput

remove_auto_scaling_policy_output

See RemoveAutoScalingPolicyOutput

remove_auto_termination_policy_output

See RemoveAutoTerminationPolicyOutput

remove_managed_scaling_policy_output

See RemoveManagedScalingPolicyOutput

remove_tags_output

See RemoveTagsOutput

run_job_flow_output

See RunJobFlowOutput

set_termination_protection_output

See SetTerminationProtectionOutput

set_visible_to_all_users_output

See SetVisibleToAllUsersOutput

start_notebook_execution_output

See StartNotebookExecutionOutput

stop_notebook_execution_output

See StopNotebookExecutionOutput

terminate_job_flows_output

See TerminateJobFlowsOutput

update_studio_output

See UpdateStudioOutput

update_studio_session_mapping_output

See UpdateStudioSessionMappingOutput

Structs

AddInstanceFleetOutput

AddInstanceGroupsOutput

Output from an AddInstanceGroups call.

AddJobFlowStepsOutput

The output for the AddJobFlowSteps operation.

AddTagsOutput

This output indicates the result of adding tags to a resource.

CancelStepsOutput

The output for the CancelSteps operation.

CreateSecurityConfigurationOutput

CreateStudioOutput

CreateStudioSessionMappingOutput

DeleteSecurityConfigurationOutput

DeleteStudioOutput

DeleteStudioSessionMappingOutput

DescribeClusterOutput

This output contains the description of the cluster.

DescribeJobFlowsOutput

The output for the DescribeJobFlows operation.

DescribeNotebookExecutionOutput

DescribeReleaseLabelOutput

DescribeSecurityConfigurationOutput

DescribeStepOutput

This output contains the description of the cluster step.

DescribeStudioOutput

GetAutoTerminationPolicyOutput

GetBlockPublicAccessConfigurationOutput

GetManagedScalingPolicyOutput

GetStudioSessionMappingOutput

ListBootstrapActionsOutput

This output contains the bootstrap actions detail.

ListClustersOutput

This contains a ClusterSummaryList with the cluster details; for example, the cluster

ListInstanceFleetsOutput

ListInstanceGroupsOutput

This input determines which instance groups to retrieve.

ListInstancesOutput

This output contains the list of instances.

ListNotebookExecutionsOutput

ListReleaseLabelsOutput

ListSecurityConfigurationsOutput

ListStepsOutput

This output contains the list of steps returned in reverse order. This means that the

ListStudioSessionMappingsOutput

ListStudiosOutput

ModifyClusterOutput

ModifyInstanceFleetOutput

ModifyInstanceGroupsOutput

PutAutoScalingPolicyOutput

PutAutoTerminationPolicyOutput

PutBlockPublicAccessConfigurationOutput

PutManagedScalingPolicyOutput

RemoveAutoScalingPolicyOutput

RemoveAutoTerminationPolicyOutput

RemoveManagedScalingPolicyOutput

RemoveTagsOutput

This output indicates the result of removing tags from the resource.

RunJobFlowOutput

The result of the RunJobFlow operation.

SetTerminationProtectionOutput

SetVisibleToAllUsersOutput

StartNotebookExecutionOutput

StopNotebookExecutionOutput

TerminateJobFlowsOutput

UpdateStudioOutput

UpdateStudioSessionMappingOutput